update deps, fix role model

+++ app/Http/Controllers/Api/PostController.php
@@ -5,11 +5,16 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
+use Spatie\Image\Manipulations;
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
+use App\Models\Category;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Exception;
 
 class PostController extends Controller
 {
+
     public function index()
     {
         $orderColumn = request('order_column', 'created_at');
@@ -20,26 +25,32 @@ class PostController extends Controller
         if (!in_array($orderDirection, ['asc', 'desc'])) {
             $orderDirection = 'desc';
         }
-        $posts = Post::with('category')
-            ->when(request('search_category'), function ($query) {
-                $query->where('category_id', request('search_category'));
+        $posts = Post::with('media')->whereHas('categories')
+            ->whereHas('categories', function ($query) {
+                if (request('search_category')) {
+                    $categories = explode(",", request('search_category'));
+                    $query->whereIn('categories.id', $categories);
+                }
             })
             ->when(request('search_id'), function ($query) {
                 $query->where('id', request('search_id'));
             })
             ->when(request('search_title'), function ($query) {
-                $query->where('title', 'like', '%'.request('search_title').'%');
+                $query->where('title', 'like', '%' . request('search_title') . '%');
             })
             ->when(request('search_content'), function ($query) {
-                $query->where('content', 'like', '%'.request('search_content').'%');
+                $query->where('content', 'like', '%' . request('search_content') . '%');
             })
             ->when(request('search_global'), function ($query) {
-                $query->where(function($q) {
+                $query->where(function ($q) {
                     $q->where('id', request('search_global'))
-                        ->orWhere('title', 'like', '%'.request('search_global').'%')
-                        ->orWhere('content', 'like', '%'.request('search_global').'%');
+                        ->orWhere('title', 'like', '%' . request('search_global') . '%')
+                        ->orWhere('content', 'like', '%' . request('search_global') . '%');
 
                 });
+            })
+            ->when(!auth()->user()->hasPermissionTo('post-all'), function ($query) {
+                $query->where('user_id', auth()->user()->id);
             })
             ->orderBy($orderColumn, $orderDirection)
             ->paginate(50);
@@ -49,57 +60,79 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $this->authorize('post-create');
-        if ($request->hasFile('thumbnail')) {
-            $filename = $request->file('thumbnail')->getClientOriginalName();
-            info($filename);
-        }
+
 
         $validatedData = $request->validated();
         $validatedData['user_id'] = auth()->id();
-
         $post = Post::create($validatedData);
 
+
+        $categories = explode(",", $request->categories);
+        $category = Category::findMany($categories);
+        $post->categories()->attach($category);
+//        try {
+        if ($request->hasFile('thumbnail')) {
+            $post->addMediaFromRequest('thumbnail')->preservingOriginal()->toMediaCollection('images');
+        }
+//        } catch (Exception $e) {
+//            error_log($e->getMessage());
+//        }
         return new PostResource($post);
     }
 
     public function show(Post $post)
     {
         $this->authorize('post-edit');
-        return new PostResource($post);
+        if ($post->user_id !== auth()->user()->id && !auth()->user()->hasPermissionTo('post-all')) {
+            return response()->json(['status' => 405, 'success' => false, 'message' => 'You can only edit your own posts']);
+        } else {
+            return new PostResource($post);
+        }
     }
 
     public function update(Post $post, StorePostRequest $request)
     {
         $this->authorize('post-edit');
-        $post->update($request->validated());
+        if ($post->user_id !== auth()->user()->id && !auth()->user()->hasPermissionTo('post-all')) {
+            return response()->json(['status' => 405, 'success' => false, 'message' => 'You can only edit your own posts']);
+        } else {
+            $post->update($request->validated());
+//            error_log(json_encode($request->categories));
 
-        return new PostResource($post);
+            $category = Category::findMany($request->categories);
+            $post->categories()->sync($category);
+            return new PostResource($post);
+        }
     }
 
-    public function destroy(Post $post) {
+    public function destroy(Post $post)
+    {
         $this->authorize('post-delete');
-        $post->delete();
-
-        return response()->noContent();
+        if ($post->user_id !== auth()->user()->id && !auth()->user()->hasPermissionTo('post-all')) {
+            return response()->json(['status' => 405, 'success' => false, 'message' => 'You can only delete your own posts']);
+        } else {
+            $post->delete();
+            return response()->noContent();
+        }
     }
 
     public function getPosts()
     {
-        $posts = Post::latest()->paginate();
+        $posts = Post::with('categories')->with('media')->latest()->paginate();
+        return PostResource::collection($posts);
 
-        return $posts;
     }
 
     public function getCategoryByPosts($id)
     {
-        $posts = Post::latest()->where('category_id', $id)->paginate();
+        $posts = Post::whereRelation('categories', 'category_id', '=', $id)->paginate();
 
-        return $posts;
+        return PostResource::collection($posts);
     }
 
     public function getPost($id)
     {
-        $post = Post::with('category', 'user')->findOrFail($id);
+        $post = Post::with('categories', 'user')->findOrFail($id);
 
         return $post;
     }
